+++ CHANGELOG.md
@@ -85,3 +85,9 @@ All notable changes to `aws-s3-helpers` will be documented in this file
 - fix return type of `S3` upload function to return the class instance instead of a url
 - optimize `S3::upload()` & `S3::upload_raw()` methods
 - refactor `S3::upload_raw()` to `S3::uploadRaw()`
+
+
+## 0.10.0 - 2021-07-01
+- optimize `s3FileUrl()` helper method by removing $temp param
+- cut `S3::list()` method as it was behaving unexpectedly and is no longer used
+- fix use of `Storage::disk('s3')` with `Storage::disk(config('filesystem.cloud', 's3'))` to make use of config